pyproject.toml
[project]
name = "llm-pydanitc-agent"
version = "0.1.0"
description = "AI Agent with MCP Server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gradio>=5.25.2",
    "logfire>=3.14.0",
    "loguru>=0.7.3",
    "mcp>=1.6.0",
    "pydantic-ai>=0.2.4",
    "python-dotenv>=1.1.0",
]
